+++ src/components/Categories/CategoriesContent.tsx
